refactor: move errors to a unified module tree
