+++ crypto-ffi/src/generic/mod.rs
@@ -12,11 +12,11 @@ use log_reload::ReloadLog;
 use tls_codec::Deserialize;
 
 use self::context::CoreCryptoContext;
-use crate::proteus_impl;
+use crate::{CoreCryptoError, CoreCryptoResult, proteus_impl};
 use core_crypto::mls::conversation::Conversation as _;
 pub use core_crypto::prelude::ConversationId;
 use core_crypto::{
-    InnermostErrorMessage, RecursiveError,
+    RecursiveError,
     prelude::{
         Client, EntropySeed, MlsBufferedConversationDecryptMessage, MlsCiphersuite, MlsClientConfiguration,
         MlsCommitBundle, MlsConversationDecryptMessage, MlsCustomConfiguration, MlsGroupInfoBundle, MlsProposalBundle,
@@ -27,308 +27,6 @@ use core_crypto_keystore::Connection as Database;
 
 pub mod context;
 mod epoch_observer;
-
-#[derive(Debug, thiserror::Error, uniffi::Error)]
-pub enum MlsError {
-    #[error("Conversation already exists")]
-    ConversationAlreadyExists(core_crypto::prelude::ConversationId),
-    #[error("We already decrypted this message once")]
-    DuplicateMessage,
-    #[error("Incoming message is for a future epoch. We will buffer it until the commit for that epoch arrives")]
-    BufferedFutureMessage,
-    #[error("Incoming message is from an epoch too far in the future to buffer.")]
-    WrongEpoch,
-    #[error(
-        "Incoming message is a commit for which we have not yet received all the proposals. Buffering until all proposals have arrived."
-    )]
-    BufferedCommit,
-    #[error("The epoch in which message was encrypted is older than allowed")]
-    MessageEpochTooOld,
-    #[error("Tried to decrypt a commit created by self which is likely to have been replayed by the DS")]
-    SelfCommitIgnored,
-    #[error(
-        "You tried to join with an external commit but did not merge it yet. We will reapply this message for you when you merge your external commit"
-    )]
-    UnmergedPendingGroup,
-    #[error("The received proposal is deemed stale and is from an older epoch.")]
-    StaleProposal,
-    #[error("The received commit is deemed stale and is from an older epoch.")]
-    StaleCommit,
-    /// This happens when the DS cannot flag KeyPackages as claimed or not. It this scenario, a client
-    /// requests their old KeyPackages to be deleted but one has already been claimed by another client to create a Welcome.
-    /// In that case the only solution is that the client receiving such a Welcome tries to join the group
-    /// with an External Commit instead
-    #[error(
-        "Although this Welcome seems valid, the local KeyPackage it references has already been deleted locally. Join this group with an external commit"
-    )]
-    OrphanWelcome,
-    /// Message rejected by the delivery service
-    #[error("Message rejected by the delivery service. Reason: {reason}")]
-    MessageRejected {
-        /// Why was the message rejected by the delivery service?
-        reason: String,
-    },
-    #[error("{0}")]
-    Other(String),
-}
-
-impl From<core_crypto::MlsError> for MlsError {
-    #[inline]
-    fn from(e: core_crypto::MlsError) -> Self {
-        Self::Other(e.innermost_error_message())
-    }
-}
-
-#[cfg(feature = "proteus")]
-#[derive(Debug, thiserror::Error, uniffi::Error)]
-pub enum ProteusError {
-    #[error("The requested session was not found")]
-    SessionNotFound,
-    #[error("We already decrypted this message once")]
-    DuplicateMessage,
-    #[error("The remote identity has changed")]
-    RemoteIdentityChanged,
-    #[error("Another Proteus error occurred but the details are probably irrelevant to clients")]
-    Other(u16),
-}
-
-#[cfg(feature = "proteus")]
-impl ProteusError {
-    pub fn from_error_code(code: u16) -> Self {
-        match code {
-            102 => Self::SessionNotFound,
-            204 => Self::RemoteIdentityChanged,
-            209 => Self::DuplicateMessage,
-            _ => Self::Other(code),
-        }
-    }
-
-    pub fn error_code(&self) -> u16 {
-        match self {
-            Self::SessionNotFound => 102,
-            Self::RemoteIdentityChanged => 204,
-            Self::DuplicateMessage => 209,
-            Self::Other(code) => *code,
-        }
-    }
-}
-
-#[cfg(feature = "proteus")]
-impl From<core_crypto::ProteusError> for ProteusError {
-    fn from(value: core_crypto::ProteusError) -> Self {
-        type SessionError = proteus_wasm::session::Error<core_crypto_keystore::CryptoKeystoreError>;
-        match value.source {
-            core_crypto::ProteusErrorKind::ProteusSessionError(SessionError::InternalError(
-                proteus_wasm::internal::types::InternalError::NoSessionForTag,
-            )) => Self::SessionNotFound,
-            core_crypto::ProteusErrorKind::ProteusSessionError(SessionError::DuplicateMessage) => {
-                Self::DuplicateMessage
-            }
-            core_crypto::ProteusErrorKind::ProteusSessionError(SessionError::RemoteIdentityChanged) => {
-                Self::RemoteIdentityChanged
-            }
-            _ => Self::Other(value.source.error_code().unwrap_or_default()),
-        }
-    }
-}
-
-#[derive(Debug, thiserror::Error, uniffi::Error)]
-pub enum CoreCryptoError {
-    #[error(transparent)]
-    Mls(#[from] MlsError),
-    #[cfg(feature = "proteus")]
-    #[error(transparent)]
-    Proteus(#[from] ProteusError),
-    #[error("End to end identity error: {0}")]
-    E2eiError(String),
-    #[error("error from client: {0}")]
-    ClientError(String),
-    #[error("{0}")]
-    Other(String),
-}
-
-/// Prepare and dispatch a log message reporting this error.
-///
-/// We want to ensure consistent logging every time we pass a log message across the FFI boundary,
-/// as we cannot guarantee the method, format, or existence of error logging once the result crosses.
-/// Unfortunately, as there is no single point at which we convert internal errors to trans-ffi
-/// errors, we need to extract the logging procedure and ensure it's called at each relevant point.
-///
-/// This has the further disadvantage that we have very little context information at the point of
-/// logging. We'll try this out for now anyway; if it turns out that we need to add more tracing
-/// in the future, we can figure out our techniques then.
-fn log_error(error: &dyn std::error::Error) {
-    // we exclude the original error message from the chain
-    let chain = {
-        let mut error = error;
-        let mut chain = Vec::new();
-        while let Some(inner) = error.source() {
-            chain.push(inner.to_string());
-            error = inner;
-        }
-        chain
-    };
-    let msg = error.to_string();
-    let err = serde_json::json!({"msg": msg, "chain": chain});
-    // even though there exists a `:err` formatter, it only captures the top-level
-    // message from the error, so it's still worth building our own inner error formatter
-    // and using serde here
-    log::warn!(target: "core-crypto", err:serde; "core-crypto returning this error across ffi; see recent log messages for context");
-}
-
-impl From<RecursiveError> for CoreCryptoError {
-    fn from(error: RecursiveError) -> Self {
-        log_error(&error);
-
-        // check if the innermost error is any kind of e2e error
-        let innermost = {
-            let mut err: &dyn std::error::Error = &error;
-            while let Some(inner) = err.source() {
-                err = inner;
-            }
-            err
-        };
-
-        if let Some(err) = innermost.downcast_ref::<core_crypto::e2e_identity::Error>() {
-            return CoreCryptoError::E2eiError(err.to_string());
-        }
-
-        // What now? We only really care about the innermost variants, not the error stack, but that produces
-        // an arbitrary set of types. We can't match against that!
-        //
-        // Or at least, not without the power of macros. We can use them to match against heterogenous types.
-
-        /// Like [`matches!`], but with an out expression which can reference items captured by the pattern.
-        ///
-        /// Hopefully only ever use this in conjunction with `interior_matches!`, because for most sane
-        /// circumstances, `if let` is the better design pattern.
-        macro_rules! matches_option {
-            ($val:expr, $pattern:pat $(if $guard:expr)? => $out:expr) => {
-                match ($val) {
-                    $pattern $(if $guard)? => Some($out),
-                    _ => None,
-                }
-            };
-        }
-
-        /// This is moderately horrific and we hopefully will not require it anywhere else, but
-        /// it solves a real problem here: how do we match against the innermost error variants,
-        /// when we have a heterogenous set of types to match against?
-        macro_rules! match_heterogenous {
-            ($err:expr => {
-                $( $pattern:pat $(if $guard:expr)? => $var:expr, )*
-                ||=> $default:expr,
-            }) => {{
-                if false {unreachable!()}
-                $(
-                    else if let Some(v) = matches_option!($err.downcast_ref(), Some($pattern) $(if $guard)? => $var) {
-                        v
-                    }
-                )*
-                else {
-                    $default
-                }
-            }};
-        }
-
-        match_heterogenous!(innermost => {
-            core_crypto::LeafError::ConversationAlreadyExists(id) => MlsError::ConversationAlreadyExists(id.clone()).into(),
-            core_crypto::mls::conversation::Error::BufferedFutureMessage{..} => MlsError::BufferedFutureMessage.into(),
-            core_crypto::mls::conversation::Error::DuplicateMessage => MlsError::DuplicateMessage.into(),
-            core_crypto::mls::conversation::Error::MessageEpochTooOld => MlsError::MessageEpochTooOld.into(),
-            core_crypto::mls::conversation::Error::SelfCommitIgnored => MlsError::SelfCommitIgnored.into(),
-            core_crypto::mls::conversation::Error::StaleCommit => MlsError::StaleCommit.into(),
-            core_crypto::mls::conversation::Error::StaleProposal => MlsError::StaleProposal.into(),
-            core_crypto::mls::conversation::Error::UnbufferedFarFutureMessage => MlsError::WrongEpoch.into(),
-            core_crypto::mls::conversation::Error::BufferedCommit => MlsError::BufferedCommit.into(),
-            core_crypto::mls::conversation::Error::MessageRejected { reason } => MlsError::MessageRejected { reason: reason.clone() }.into(),
-            core_crypto::mls::conversation::Error::OrphanWelcome => MlsError::OrphanWelcome.into(),
-            // The internal name is what we want, but renaming the external variant is a breaking change.
-            // Since we're re-designing the `BufferedMessage` errors soon, it's not worth producing
-            // an additional breaking change until then, so the names are inconsistent.
-            core_crypto::mls::conversation::Error::BufferedForPendingConversation => MlsError::UnmergedPendingGroup.into(),
-            ||=> MlsError::Other(error.innermost_error_message()).into(),
-        })
-    }
-}
-
-// This implementation is intended to be temporary; we're going to be completely restructuring the way we handle
-// errors in `core-crypto` soon. We can replace this with better error patterns when we do.
-//
-// Certain error mappings could apply to both MLS and Proteus. In all such cases, we map them to the MLS variant.
-// When we redesign the errors in `core-crypto`, these ambiguities should disappear anyway.
-impl From<core_crypto::Error> for CoreCryptoError {
-    fn from(error: core_crypto::Error) -> Self {
-        log_error(&error);
-
-        // we can take care of the _simple_ error-mapping up here.
-        #[cfg(feature = "proteus")]
-        if let core_crypto::Error::Proteus(proteus) = &error {
-            if let Some(code) = proteus.source.error_code() {
-                if code != 0 {
-                    return Self::Proteus(ProteusError::from_error_code(code));
-                }
-            }
-        }
-        match error {
-            core_crypto::Error::ProteusNotInitialized => Self::Other("proteus not initialized".to_string()),
-            core_crypto::Error::Proteus(proteus) => Self::Other(proteus.innermost_error_message()),
-            core_crypto::Error::Mls(mls) => Self::Mls(MlsError::from(mls)),
-            core_crypto::Error::InvalidContext => Self::Other(error.to_string()),
-            core_crypto::Error::MlsTransportNotProvided => Self::Other(error.to_string()),
-            core_crypto::Error::ErrorDuringMlsTransport(error_message) => Self::Other(error_message),
-            core_crypto::Error::Keystore(keystore_error) => Self::Other(keystore_error.innermost_error_message()),
-            core_crypto::Error::CryptoboxMigration(cryptobox) => Self::Other(cryptobox.innermost_error_message()),
-            core_crypto::Error::Recursive(recursive_error) => recursive_error.into(),
-            core_crypto::Error::FeatureDisabled(_) => Self::Other(error.to_string()),
-        }
-    }
-}
-
-/// We can't do a generic `impl<E: ToRecursiveError> From<E> for CoreCryptoError`
-/// because that has the potential to cause breaking conflicts later on: what if
-/// core-crypto later did `impl ToRecursiveError for core_crypto::Error`? That would
-/// cause a duplicate `From` impl.
-///
-/// Instead, we explicitly specify every variant which can be converted to a
-/// `CoreCryptoError`, and implement its `From` block directly.
-macro_rules! impl_from_via_recursive_error {
-    ($($t:ty),+ $(,)?) => {
-        $(
-            impl From<$t> for CoreCryptoError {
-                fn from(error: $t) -> Self {
-                    use core_crypto::ToRecursiveError;
-                    error
-                        .construct_recursive("this context string does not matter and gets immediately stripped")
-                        .into()
-                }
-            }
-        )*
-    };
-}
-
-impl_from_via_recursive_error!(
-    core_crypto::mls::Error,
-    core_crypto::mls::conversation::Error,
-    core_crypto::e2e_identity::Error,
-);
-
-impl From<uniffi::UnexpectedUniFFICallbackError> for CoreCryptoError {
-    fn from(value: uniffi::UnexpectedUniFFICallbackError) -> Self {
-        Self::ClientError(value.reason)
-    }
-}
-
-impl CoreCryptoError {
-    fn generic<E>() -> impl FnOnce(E) -> Self
-    where
-        E: ToString,
-    {
-        |err| Self::Other(err.to_string())
-    }
-}
-
-type CoreCryptoResult<T> = Result<T, CoreCryptoError>;
 
 #[derive(Debug, Clone, Eq, Hash, PartialEq, derive_more::From)]
 pub struct ClientId(core_crypto::prelude::ClientId);
@@ -1579,6 +1277,8 @@ impl From<AcmeChallenge> for core_crypto::prelude::E2eiAcmeChallenge {
 
 #[cfg(test)]
 mod tests {
+    use crate::MlsError;
+
     use super::*;
     use core_crypto::LeafError;
     #[test]
