+++ crypto-ffi/src/wasm/mod.rs
@@ -11,7 +11,7 @@ use std::{
 
 use crate::proteus_impl;
 use core_crypto::mls::conversation::Conversation as _;
-use core_crypto::{InnermostErrorMessage, MlsTransportResponse, prelude::*};
+use core_crypto::{MlsTransportResponse, prelude::*};
 use core_crypto_keystore::Connection as Database;
 use futures_util::future::TryFutureExt;
 use js_sys::{Promise, Uint8Array};
@@ -25,391 +25,7 @@ use utils::*;
 use wasm_bindgen::{JsCast, prelude::*};
 use wasm_bindgen_futures::future_to_promise;
 
-#[derive(Debug, thiserror::Error, strum::AsRefStr)]
-pub enum MlsError {
-    #[error("Conversation already exists")]
-    ConversationAlreadyExists(core_crypto::prelude::ConversationId),
-    #[error("We already decrypted this message once")]
-    DuplicateMessage,
-    #[error("Incoming message is for a future epoch. We will buffer it until the commit for that epoch arrives")]
-    BufferedFutureMessage,
-    #[error("Incoming message is from an epoch too far in the future to buffer.")]
-    WrongEpoch,
-    #[error(
-        "Incoming message is a commit for which we have not yet received all the proposals. Buffering until all proposals have arrived."
-    )]
-    BufferedCommit,
-    #[error("The epoch in which message was encrypted is older than allowed")]
-    MessageEpochTooOld,
-    #[error("Tried to decrypt a commit created by self which is likely to have been replayed by the DS")]
-    SelfCommitIgnored,
-    #[error(
-        "You tried to join with an external commit but did not merge it yet. We will reapply this message for you when you merge your external commit"
-    )]
-    UnmergedPendingGroup,
-    #[error("The received proposal is deemed stale and is from an older epoch.")]
-    StaleProposal,
-    #[error("The received commit is deemed stale and is from an older epoch.")]
-    StaleCommit,
-    /// This happens when the DS cannot flag KeyPackages as claimed or not. It this scenario, a client
-    /// requests their old KeyPackages to be deleted but one has already been claimed by another client to create a Welcome.
-    /// In that case the only solution is that the client receiving such a Welcome tries to join the group
-    /// with an External Commit instead
-    #[error(
-        "Although this Welcome seems valid, the local KeyPackage it references has already been deleted locally. Join this group with an external commit"
-    )]
-    OrphanWelcome,
-    /// Message rejected by the delivery service
-    #[error("Message rejected by the delivery service. Reason: {reason}")]
-    MessageRejected {
-        /// Why was the message rejected by the delivery service?
-        reason: String,
-    },
-    #[error("{0}")]
-    Other(String),
-}
-
-impl From<core_crypto::MlsError> for MlsError {
-    #[inline]
-    fn from(e: core_crypto::MlsError) -> Self {
-        Self::Other(e.innermost_error_message())
-    }
-}
-
-#[cfg(feature = "proteus")]
-#[derive(Debug, thiserror::Error, strum::AsRefStr)]
-pub enum ProteusError {
-    #[error("Proteus was not initialized")]
-    NotInitialized,
-    #[error("The requested session was not found")]
-    SessionNotFound,
-    #[error("We already decrypted this message once")]
-    DuplicateMessage,
-    #[error("The remote identity has changed")]
-    RemoteIdentityChanged,
-    #[error("Another Proteus error occurred but the details are probably irrelevant to clients ({})", .0.unwrap_or_default())]
-    Other(Option<u16>),
-}
-
-impl ProteusError {
-    pub fn from_error_code(code: impl Into<Option<u16>>) -> Option<Self> {
-        let code = code.into()?;
-        if code == 0 {
-            return None;
-        }
-
-        match code {
-            5 => Self::NotInitialized,
-            102 => Self::SessionNotFound,
-            204 => Self::RemoteIdentityChanged,
-            209 => Self::DuplicateMessage,
-            _ => Self::Other(Some(code)),
-        }
-        .into()
-    }
-
-    pub fn error_code(&self) -> Option<u16> {
-        match self {
-            ProteusError::NotInitialized => Some(5),
-            ProteusError::SessionNotFound => Some(102),
-            ProteusError::RemoteIdentityChanged => Some(204),
-            ProteusError::DuplicateMessage => Some(209),
-            ProteusError::Other(code) => *code,
-        }
-    }
-}
-
-#[cfg(feature = "proteus")]
-impl From<core_crypto::ProteusError> for ProteusError {
-    fn from(value: core_crypto::ProteusError) -> Self {
-        type SessionError = proteus_wasm::session::Error<core_crypto_keystore::CryptoKeystoreError>;
-        match value.source {
-            core_crypto::ProteusErrorKind::ProteusSessionError(SessionError::InternalError(
-                proteus_wasm::internal::types::InternalError::NoSessionForTag,
-            )) => Self::SessionNotFound,
-            core_crypto::ProteusErrorKind::ProteusSessionError(SessionError::DuplicateMessage) => {
-                Self::DuplicateMessage
-            }
-            core_crypto::ProteusErrorKind::ProteusSessionError(SessionError::RemoteIdentityChanged) => {
-                Self::RemoteIdentityChanged
-            }
-            _ => Self::Other(value.source.error_code()),
-        }
-    }
-}
-
-#[derive(Debug, thiserror::Error, strum::AsRefStr)]
-pub(crate) enum InternalError {
-    #[error(transparent)]
-    MlsError(#[from] MlsError),
-    #[cfg(feature = "proteus")]
-    #[error(transparent)]
-    ProteusError(#[from] ProteusError),
-    #[error("End to end identity error: {0}")]
-    E2eiError(String),
-    #[error(transparent)]
-    SerializationError(#[from] serde_wasm_bindgen::Error),
-    #[error("Unknown ciphersuite identifier")]
-    UnknownCiphersuite,
-    #[error("Transaction rolled back. Uncaught JsError: {uncaught_error:?}")]
-    TransactionFailed { uncaught_error: JsValue },
-    #[error("{0}")]
-    Other(String),
-}
-
-/// Prepare and dispatch a log message reporting this error.
-///
-/// We want to ensure consistent logging every time we pass a log message across the FFI boundary,
-/// as we cannot guarantee the method, format, or existence of error logging once the result crosses.
-/// In this case there is a single point at which we convert internal errors to trans-ffi
-/// errors, but it was still convenient to extract the logging procedure, because that point is
-/// within a macro-generated `From` impl.
-///
-/// This has the further disadvantage that we have very little context information at the point of
-/// logging. We'll try this out for now anyway; if it turns out that we need to add more tracing
-/// in the future, we can figure out our techniques then.
-fn log_error(error: &dyn std::error::Error) {
-    // we exclude the original error message from the chain
-    let chain = {
-        let mut error = error;
-        let mut chain = Vec::new();
-        while let Some(inner) = error.source() {
-            chain.push(inner.to_string());
-            error = inner;
-        }
-        chain
-    };
-    let msg = error.to_string();
-    let err = serde_json::json!({"msg": msg, "chain": chain});
-    // even though there exists a `:err` formatter, it only captures the top-level
-    // message from the error, so it's still worth building our own inner error formatter
-    // and using serde here
-    log::warn!(target: "core-crypto", err:serde; "core-crypto returning this error across ffi; see recent log messages for context");
-}
-
-impl From<RecursiveError> for InternalError {
-    fn from(error: RecursiveError) -> Self {
-        log_error(&error);
-
-        // check if the innermost error is any kind of e2e error
-        let innermost = {
-            let mut err: &dyn std::error::Error = &error;
-            while let Some(inner) = err.source() {
-                err = inner;
-            }
-            err
-        };
-
-        if let Some(err) = innermost.downcast_ref::<core_crypto::e2e_identity::Error>() {
-            return InternalError::E2eiError(err.to_string());
-        }
-
-        // What now? We only really care about the innermost variants, not the error stack, but that produces
-        // an arbitrary set of types. We can't match against that!
-        //
-        // Or at least, not without the power of macros. We can use them to match against heterogenous types.
-
-        /// Like [`matches!`], but with an out expression which can reference items captured by the pattern.
-        ///
-        /// Hopefully only ever use this in conjunction with `interior_matches!`, because for most sane
-        /// circumstances, `if let` is the better design pattern.
-        macro_rules! matches_option {
-            ($val:expr, $pattern:pat $(if $guard:expr)? => $out:expr) => {
-                match ($val) {
-                    $pattern $(if $guard)? => Some($out),
-                    _ => None,
-                }
-            };
-        }
-
-        /// This is moderately horrific and we hopefully will not require it anywhere else, but
-        /// it solves a real problem here: how do we match against the innermost error variants,
-        /// when we have a heterogenous set of types to match against?
-        macro_rules! match_heterogenous {
-            ($err:expr => {
-                $( $pattern:pat $(if $guard:expr)? => $var:expr, )*
-                ||=> $default:expr,
-            }) => {{
-                if false {unreachable!()}
-                $(
-                    else if let Some(v) = matches_option!($err.downcast_ref(), Some($pattern) $(if $guard)? => $var) {
-                        v
-                    }
-                )*
-                else {
-                    $default
-                }
-            }};
-        }
-
-        match_heterogenous!(innermost => {
-            core_crypto::LeafError::ConversationAlreadyExists(id) => MlsError::ConversationAlreadyExists(id.clone()).into(),
-            core_crypto::mls::conversation::Error::BufferedFutureMessage{..} => MlsError::BufferedFutureMessage.into(),
-            core_crypto::mls::conversation::Error::DuplicateMessage => MlsError::DuplicateMessage.into(),
-            core_crypto::mls::conversation::Error::MessageEpochTooOld => MlsError::MessageEpochTooOld.into(),
-            core_crypto::mls::conversation::Error::SelfCommitIgnored => MlsError::SelfCommitIgnored.into(),
-            core_crypto::mls::conversation::Error::StaleCommit => MlsError::StaleCommit.into(),
-            core_crypto::mls::conversation::Error::StaleProposal => MlsError::StaleProposal.into(),
-            core_crypto::mls::conversation::Error::UnbufferedFarFutureMessage => MlsError::WrongEpoch.into(),
-            core_crypto::mls::conversation::Error::BufferedCommit => MlsError::BufferedCommit.into(),
-            core_crypto::mls::conversation::Error::MessageRejected { reason } => MlsError::MessageRejected { reason: reason.clone() }.into(),
-            core_crypto::mls::conversation::Error::OrphanWelcome => MlsError::OrphanWelcome.into(),
-            // The internal name is what we want, but renaming the external variant is a breaking change.
-            // Since we're re-designing the `BufferedMessage` errors soon, it's not worth producing
-            // an additional breaking change until then, so the names are inconsistent.
-            core_crypto::mls::conversation::Error::BufferedForPendingConversation => MlsError::UnmergedPendingGroup.into(),
-            ||=> MlsError::Other(error.innermost_error_message()).into(),
-        })
-    }
-}
-
-// This implementation is intended to be temporary; we're going to be completely restructuring the way we handle
-// errors in `core-crypto` soon. We can replace this with better error patterns when we do.
-//
-// Certain error mappings could apply to both MLS and Proteus. In all such cases, we map them to the MLS variant.
-// When we redesign the errors in `core-crypto`, these ambiguities should disappear anyway.
-impl From<core_crypto::Error> for InternalError {
-    fn from(error: core_crypto::Error) -> Self {
-        log_error(&error);
-
-        // we can take care of the _simple_ error-mapping up here.
-        #[cfg(feature = "proteus")]
-        if let core_crypto::Error::Proteus(proteus) = &error {
-            if let Some(code) = proteus.source.error_code() {
-                if code != 0 {
-                    if let Some(proteus_error) = ProteusError::from_error_code(code) {
-                        return Self::ProteusError(proteus_error);
-                    }
-                }
-            }
-        }
-        match error {
-            #[cfg(feature = "proteus")]
-            core_crypto::Error::ProteusNotInitialized => Self::ProteusError(ProteusError::NotInitialized),
-            core_crypto::Error::Proteus(proteus) => Self::Other(proteus.innermost_error_message()),
-            core_crypto::Error::Mls(mls) => Self::MlsError(MlsError::from(mls)),
-            core_crypto::Error::InvalidContext => Self::Other(error.to_string()),
-            core_crypto::Error::MlsTransportNotProvided => Self::Other(error.to_string()),
-            core_crypto::Error::ErrorDuringMlsTransport(error_message) => Self::Other(error_message),
-            core_crypto::Error::Keystore(keystore_error) => Self::Other(keystore_error.innermost_error_message()),
-            core_crypto::Error::CryptoboxMigration(cryptobox) => Self::Other(cryptobox.innermost_error_message()),
-            core_crypto::Error::Recursive(recursive_error) => recursive_error.into(),
-            core_crypto::Error::FeatureDisabled(_) => Self::Other(error.to_string()),
-        }
-    }
-}
-
-/// We can't do a generic `impl<E: ToRecursiveError> From<E> for InternalError`
-/// because that has the potential to cause breaking conflicts later on: what if
-/// core-crypto later did `impl ToRecursiveError for core_crypto::Error`? That would
-/// cause a duplicate `From` impl.
-///
-/// Instead, we explicitly specify every variant which can be converted to a
-/// `InternalError`, and implement its `From` block directly.
-macro_rules! impl_from_via_recursive_error {
-    ($($t:ty),+ $(,)?) => {
-        $(
-            impl From<$t> for InternalError {
-                fn from(error: $t) -> Self {
-                    use core_crypto::ToRecursiveError;
-                    error
-                        .construct_recursive("this context string does not matter and gets immediately stripped")
-                        .into()
-                }
-            }
-        )*
-    };
-}
-
-impl_from_via_recursive_error!(
-    core_crypto::mls::Error,
-    core_crypto::mls::conversation::Error,
-    core_crypto::e2e_identity::Error,
-);
-
-impl InternalError {
-    fn generic<E>() -> impl FnOnce(E) -> Self
-    where
-        E: ToString,
-    {
-        |err| Self::Other(err.to_string())
-    }
-
-    fn variant_name(&self) -> String {
-        let mut out = self.as_ref().to_string();
-        match self {
-            Self::MlsError(mls) => out += mls.as_ref(),
-            Self::ProteusError(proteus) => out += proteus.as_ref(),
-            _ => {}
-        }
-        out
-    }
-
-    fn stack(&self) -> Vec<String> {
-        let mut stack = Vec::new();
-        let mut err: &dyn std::error::Error = self;
-        stack.push(err.to_string());
-
-        while let Some(source) = err.source() {
-            stack.push(source.to_string());
-            err = source;
-        }
-
-        stack
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub struct CoreCryptoError(#[source] InternalError);
-
-impl std::fmt::Display for CoreCryptoError {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        let proteus_error_code = match &self.0 {
-            InternalError::ProteusError(ProteusError::Other(code)) => *code,
-            _ => None,
-        };
-
-        let json = serde_json::to_string(&serde_json::json!({
-            "message": self.0.to_string(),
-            "error_name": self.0.variant_name(),
-            "error_stack": self.0.stack(),
-            "proteus_error_code": proteus_error_code,
-        }))
-        .map_err(|_| std::fmt::Error)?;
-
-        write!(f, "{json}")
-    }
-}
-
-impl<T> From<T> for CoreCryptoError
-where
-    T: Into<InternalError>,
-{
-    fn from(value: T) -> Self {
-        Self(value.into())
-    }
-}
-
-impl From<CoreCryptoError> for wasm_bindgen::JsValue {
-    fn from(val: CoreCryptoError) -> Self {
-        fn construct_error_stack(err: &dyn std::error::Error) -> js_sys::Error {
-            let out = js_sys::Error::new(&err.to_string());
-            if let Some(source) = err.source() {
-                let source_value = construct_error_stack(source);
-                out.set_cause(&source_value);
-            }
-            out
-        }
-
-        let stacked_error = construct_error_stack(&val);
-        stacked_error.set_name(&val.0.variant_name());
-
-        stacked_error.into()
-    }
-}
-
-pub type CoreCryptoResult<T> = Result<T, CoreCryptoError>;
-pub type WasmCryptoResult<T> = CoreCryptoResult<T>;
+use crate::{CoreCryptoError, InternalError, MlsError, WasmCryptoResult};
 
 #[allow(non_camel_case_types)]
 #[wasm_bindgen]
